+++ prompt_styles/human.sh
@@ -26,6 +26,7 @@ fancygit_prompt_builder() {
     local host="${light_yellow}\h${none}"
     local where="${light_green}\w${none}"
     local venv=""
+    local user_at_host=""
 
     if ! fg_show_full_path
     then
@@ -40,7 +41,12 @@ fancygit_prompt_builder() {
         branch_name="on ${light_magenta}$branch_name${none}"
     fi
 
-    PS1="${bold}${venv}${user} at ${host} in $where $branch_name$(fg_branch_status 1)${bold_none}\n\$ "
+    if fg_show_user_at_machine
+    then
+        user_at_host="${user} at ${host} in "
+    fi
+
+    PS1="${bold}${venv}${user_at_host}$where $branch_name$(fg_branch_status 1)${bold_none}\n\$ "
 }
 
 PROMPT_COMMAND="fancygit_prompt_builder"